requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
requests==2.32.3
Django==5.0.6
python-dotenv
replicate